CPR number must have valid format
